requirements.txt
fastapi
uvicorn
pydantic
uuid
starlette
python-dotenv
requests